[7] - Tests de casos de exio y error creados
